+++ src/main/java/roomescape/persistence/ReservationTimeRepository.java
@@ -1,7 +1,8 @@
-package roomescape.domain;
+package roomescape.persistence;
 
 import java.util.List;
 import java.util.Optional;
+import roomescape.domain.ReservationTime;
 
 public interface ReservationTimeRepository {
