+++ src/main/java/roomescape/persistence/ThemeRepository.java
@@ -1,8 +1,9 @@
-package roomescape.domain;
+package roomescape.persistence;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import roomescape.domain.Theme;
 
 public interface ThemeRepository {
