+++ src/main/java/roomescape/persistence/impl/JdbcReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.persistence;
+package roomescape.persistence.impl;
 
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -7,6 +7,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.*;
+import roomescape.persistence.ReservationRepository;
 import roomescape.persistence.query.CreateReservationQuery;
 
 import java.sql.PreparedStatement;
