+++ src/main/java/roomescape/service/ReservationService.java
@@ -3,6 +3,10 @@ package roomescape.service;
 import org.springframework.stereotype.Service;
 import roomescape.domain.*;
 import roomescape.exception.*;
+import roomescape.persistence.MemberRepository;
+import roomescape.persistence.ReservationRepository;
+import roomescape.persistence.ReservationTimeRepository;
+import roomescape.persistence.ThemeRepository;
 import roomescape.persistence.query.CreateReservationQuery;
 import roomescape.service.param.CreateReservationParam;
 import roomescape.service.result.ReservationResult;
